FEAT: Add support for validity of class Model
This commit modifies module "model.py" and " __init__.py".
In "model.py", the class Model is modified so that now cortex and
forward are no longer optional arguments, but obligatory. In addition,
a check is added where it is verifired if cortex and the forward operator
are of appropriate type, as well as the consistency between them.
Function minimal_mesh is also added which enables creating a simple mesh.
In __init__.py a function loading_models is changed. It enables
loading the files from the models directory by calling the load_model
function from "model.py".
Test for the function load_models is also added.